통합 테스트 작성 22-12-12 16:23 (김민섭)
